add static_cast 12.24 #18

+++ 12-coding-practice-problems/12.24-loops-academic-index-score-table/trang_solution.cpp
@@ -9,7 +9,7 @@ int main() {
 
     for (int i = 0; i < 11; i++) {
         GPA += 0.1;
-        test_score = ((threshold - (2.5 * (GPA / 4.0) * 100)) / 100) * 1600;
+        test_score = ((static_cast<double>(threshold) - (2.5 * (GPA / 4.0) * 100.0)) / 100.0) * 1600.0;
         cout << GPA << " " << test_score << endl;
     }
